Extract this stuff out to its own controller.
