+++ src/Http/Controllers/ReportController.php
@@ -3,8 +3,6 @@
 namespace Statamic\SeoPro\Http\Controllers;
 
 use Illuminate\Http\Request;
-use Statamic\CP\Column;
-use Statamic\Extensions\Pagination\LengthAwarePaginator;
 use Statamic\Facades\User;
 use Statamic\Http\Controllers\CP\CpController;
 use Statamic\SeoPro\Reporting\Report;
@@ -37,10 +35,6 @@ class ReportController extends CpController
 
         $report = Report::find($id);
 
-        if ($request->ajax()) {
-            return $this->reportData($report);
-        }
-
         return view('seo-pro::reports.show', ['report' => $report]);
     }
 
@@ -49,57 +43,5 @@ class ReportController extends CpController
         abort_unless(User::current()->can('delete seo reports'), 403);
 
         return Report::find($id)->delete();
-    }
-
-    public function reportData($report)
-    {
-        $data = $report->data();
-
-        $data['columns'] = [
-            Column::make('status')->label(__('Status')),
-            Column::make('url')->label(__('URL')),
-            Column::make('actionable')->label(__('Actionable'))->sortable(false),
-        ];
-
-        $data['sortColumn'] = request()->input('sortColumn', 'status');
-        $data['sortDirection'] = request()->input('sortDirection', 'asc');
-
-        $pages = $data['pages']->sortBy(
-            callback: fn ($value) => $this->sortablePageValue($value, $data['sortColumn']),
-            descending: $data['sortDirection'] === 'desc',
-        )->values();
-
-        ray($pages);
-
-        $currentPage = request()->input('page', 1);
-        $perPage = request()->input('perPage', config('statamic.cp.pagination_size')); // TODO: default to config
-
-        $data['pages'] = new LengthAwarePaginator(
-            $pages->forPage($currentPage, $perPage)->values(),
-            $pages->count(),
-            $perPage,
-            $currentPage,
-        );
-
-        return $data;
-    }
-
-    private function sortablePageValue($value, $column)
-    {
-        $value = $value[$column];
-
-        if ($column !== 'status') {
-            return $value;
-        }
-
-        if ($value === 'fail') {
-            return '1fail';
-        } elseif ($value === 'warning') {
-            return '2warning';
-        } elseif ($value === 'pass') {
-            return '3pass';
-        }
-
-        return $value;
     }
 }
